I would like to make this commit a formal apology
This might actually be some of the worst code that I have ever written and I'm ashamed to even have commit it at all. The fact that I determined teh numbers in the way I did makes me want to vomit. If you are reading this code, please just ignore this commit. I was stubborn. I went down a path that should have been determined non-viable, but decided to commit to it and I did. This is digusting, innefficient, and horrid to the naked eye. Forgive me fater for I have sinned